fix: add a new story icon, extract a method for widget capturing to utils
